requirements.txt
# Core pipeline dependencies
duckdb
polars
sodapy
python-dotenv
requests

# Orchestration
apache-airflow==2.10.3
